+++ hutool-crypto/src/test/java/cn/hutool/crypto/test/PemUtilTest.java
@@ -1,40 +1,39 @@
 package cn.hutool.crypto.test;
 
-import java.security.PrivateKey;
-import java.security.PublicKey;
-
+import cn.hutool.core.io.resource.ResourceUtil;
+import cn.hutool.crypto.PemUtil;
+import cn.hutool.crypto.asymmetric.KeyType;
+import cn.hutool.crypto.asymmetric.RSA;
 import org.junit.Assert;
 import org.junit.Test;
 
-import cn.hutool.core.io.resource.ResourceUtil;
-import cn.hutool.crypto.BCUtil;
-import cn.hutool.crypto.asymmetric.KeyType;
-import cn.hutool.crypto.asymmetric.RSA;
+import java.security.PrivateKey;
+import java.security.PublicKey;
 
-public class BCUtilTest {
-	
+public class PemUtilTest {
+
 	@Test
 	public void readPrivateKeyTest() {
-		PrivateKey privateKey = BCUtil.readPrivateKey(ResourceUtil.getStream("test_private_key.pem"));
+		PrivateKey privateKey = PemUtil.readPemPrivateKey(ResourceUtil.getStream("test_private_key.pem"));
 		Assert.assertNotNull(privateKey);
 	}
 	
 	@Test
 	public void readPublicKeyTest() {
-		PublicKey publicKey = BCUtil.readPublicKey(ResourceUtil.getStream("test_public_key.csr"));
+		PublicKey publicKey = PemUtil.readPemPublicKey(ResourceUtil.getStream("test_public_key.csr"));
 		Assert.assertNotNull(publicKey);
 	}
 
 	@Test
 	public void readPemKeyTest() {
-		PublicKey publicKey = (PublicKey) BCUtil.readPemKey(ResourceUtil.getStream("test_public_key.csr"));
+		PublicKey publicKey = (PublicKey) PemUtil.readPemKey(ResourceUtil.getStream("test_public_key.csr"));
 		Assert.assertNotNull(publicKey);
 	}
 
 	@Test
 	public void validateKey() {
-		PrivateKey privateKey = BCUtil.readPrivateKey(ResourceUtil.getStream("test_private_key.pem"));
-		PublicKey publicKey = BCUtil.readPublicKey(ResourceUtil.getStream("test_public_key.csr"));
+		PrivateKey privateKey = PemUtil.readPemPrivateKey(ResourceUtil.getStream("test_private_key.pem"));
+		PublicKey publicKey = PemUtil.readPemPublicKey(ResourceUtil.getStream("test_public_key.csr"));
 		
 		RSA rsa = new RSA(privateKey, publicKey);
 		String str = "你好，Hutool";//测试字符串